Add debug mode

+++ build_and_run.sh
@@ -5,12 +5,14 @@ GIT_BRANCH="feat_add_quickwit_output"
 
 [[ ! -d falco_src ]] && git clone "${GIT_REPO}" falco_src
 
-cd falco_src
-git checkout "${GIT_BRANCH}"
-git add .
-git stash
-git stash clear
-git pull --rebase
-cd -
+if [[ $1 != "debug" ]]; then
+    cd falco_src
+    git checkout "${GIT_BRANCH}"
+    git add .
+    git stash
+    git stash clear
+    git pull --rebase
+    cd -
+fi
 
 docker-compose up --build --force-recreate
